WIP: Add image cache metadata
This adds a mechanism to store additional metadata alongside cached images.  The first use of this metadata cache is to store the Content-Type of the image so that we don't have to rely on making assumptions based on the filename, but can instead use the ACTUAL image format information obtained when originally processing the image.