Add ImageMetadata
Refactor IImageResolver so that the LastModified timestamp and the ContentType string are exposed via a new ImageMetadata class that can be asynchronously retrieved from a resolver.  Similarly, when storing an image into an IImageCache, an ImageMetadata object is supplied along with the raw image stream.

As part of this change, the logic for handling expired cache entries was shuffled around a little bit.  The ImageSharpMiddleware now handles the logic for determining if the image available from the cache is still valid for use or not, and the IImageCache simply acts as a factory for IImageResolver instances.