WIP: Add image cache metadata (#50)
* WIP: Add image cache metadata

This adds a mechanism to store additional metadata alongside cached images.  The first use of this metadata cache is to store the Content-Type of the image so that we don't have to rely on making assumptions based on the filename, but can instead use the ACTUAL image format information obtained when originally processing the image.

* Add ImageMetadata

Refactor IImageResolver so that the LastModified timestamp and the ContentType string are exposed via a new ImageMetadata class that can be asynchronously retrieved from a resolver.  Similarly, when storing an image into an IImageCache, an ImageMetadata object is supplied along with the raw image stream.

As part of this change, the logic for handling expired cache entries was shuffled around a little bit.  The ImageSharpMiddleware now handles the logic for determining if the image available from the cache is still valid for use or not, and the IImageCache simply acts as a factory for IImageResolver instances.

* Cleanup + use bytes for smallest possible meta file.

* Cache now generates extension based on correct content type.

* Try removing inlining from Parse

* Use correct string size and reset stream.

* Use string based metadata

* Allow variable Chache-Control Max-Age