Fix race condition in AsyncKeyLock
Add a RefCount property to the Doorman class, and then use a SpinLock in AsyncKeyLock to ensure that retrieving/removing a Doorman from the Keys dictionary happens atomically with the incrementing/decrementing of that Doorman's RefCount property.

Also fixed a bug in Doorman where the status field was not getting reset to 0 when releasing the final writer lock.